# docs/api/openapi.yaml
openapi: 3.0.1
info:
  title: EasyConf Reviews Microservice
  description: "EasyConf is web-based conference management system, designed to organize\
    \ and administrate academic conferences, workshops, and events. It offers a platform\
    \ for authors, reviewers, and organizers (chairs) to handle paper submissions,\
    \ peer reviews, and program scheduling efficiently. It provides tools for managing\
    \ the review process, generating program schedules, and handling communication\
    \ between organizers and participants.\n\nThe Reviews Microservice handles the\
    \ process of reviewing papers before a conference. This microservice is responsible\
    \ for assigning papers to reviews, allowing reviewers to add their comments, and\
    \ allowing the PC chairs to finalize the decision about acceptance and rejection\
    \ of the submitted papers.\n\nThe different roles that the users can have are\
    \ specified in the Users Microservice and the process for submitting papers to\
    \ conferences is specified in the Submissions Microservice.\n\nTECHNICAL NOTE:\
    \ in every endpoint, we require the requesting user to provide their ID. I.e.,\
    \ the requesterID parameter must be provided. It is used for authentication in\
    \ order to check if the user is allowed to perform certain actions."
  version: 0.0.1-SNAPSHOT
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /papers/{paperID}/reviews/finalization:
    post:
      tags:
      - Reviews
      summary: End the discussion phase for a paper
      description: "The paper is now considered reviewed: the reviews cannot be changed.\
        \ Once all of the papers in a track are reviewed, the track automatically\
        \ goes to the FINAL phase."
      operationId: finalization
      parameters:
      - name: requesterID
        in: query
        description: The ID of a user making the request
        required: true
        schema:
          type: integer
          format: int64
      - name: paperID
        in: path
        description: The ID of a paper to finalize the reviews for
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "409":
          description: Conflict. The paper must be in its Discussion phase and all
            of the reviews need to either be positive or negative. Only then can they
            be finalized.
        "500":
          description: Internal Server Error. An unexpected server error occurred.
        "200":
          description: Successful finalization of the discussion phase
        "403":
          description: Forbidden. The requester must be a chair of the track the paper
            is in.
        "404":
          description: Not Found. The requested paper was not found.
  /papers/{paperID}/reviews/by-reviewer/{reviewerID}/discussion-comments:
    get:
      tags:
      - Reviews
      summary: Gets the discussion comments
      description: "Responds with all the discussion comments for a paper. Discussion\
        \ comments are comments that can be left on reviews during the Discussion\
        \ phase. These comments will not be revealed to authors. The requester must\
        \ be a chair of the track that the paper is in, or a reviewer also assigned\
        \ to the given paper. Once posted, these comments cannot be edited or deleted."
      operationId: getDiscussionComments
      parameters:
      - name: requesterID
        in: query
        description: The ID of the user making the request.
        required: true
        schema:
          type: integer
          format: int64
      - name: reviewerID
        in: path
        description: The ID of a reviewer in charge of this paper
        required: true
        schema:
          type: integer
          format: int64
      - name: paperID
        in: path
        description: The ID of a paper to get the comments from
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DiscussionComment'
        "500":
          description: Internal Server Error. An unexpected server error occurred.
        "404":
          description: Not Found. The requested paper or reviewer was not found.
        "403":
          description: Forbidden. The requester is not a valid chair or reviewer.
    post:
      tags:
      - Reviews
      summary: Posts a discussion comment
      description: "Posts a discussion comment for a review of a specific paper using\
        \ userID and reviewerID. Discussion comments are comments that can be left\
        \ on reviews during the Discussion phase. These comments will not be revealed\
        \ to authors. The requester must be a chair of the track that the paper is\
        \ in, or a reviewer also assigned to the given paper. Once posted, these comments\
        \ cannot be edited or deleted."
      operationId: submitConfidentialComment
      parameters:
      - name: requesterID
        in: query
        description: The ID of a user making the request
        required: true
        schema:
          type: integer
          format: int64
      - name: reviewerID
        in: path
        description: The ID of a reviewer in charge of this paper
        required: true
        schema:
          type: integer
          format: int64
      - name: paperID
        in: path
        description: The ID of a paper to which to add the comment to
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: string
              description: Comment
              example: Some comment
        required: true
      responses:
        "500":
          description: Internal Server Error. An unexpected server error occurred.
        "201":
          description: Success. A discussion comment was appended to the review.
        "404":
          description: Not Found. The requested paper or reviewer was not found.
        "403":
          description: Forbidden. The requester is not a valid chair or reviewer.
...
  /conferences/{conferenceID}/tracks/{trackID}/finalization:
    post:
      tags:
      - Assignments
      summary: Finalise reviewers
      description: "Finalises the assignment of reviewers, so they can no longer be\
        \ changed manually or automatically. "
      operationId: finalization_1
      parameters:
      - name: requesterID
        in: query
        description: The ID of the user making the request
        required: true
        schema:
          type: integer
          format: int64
      - name: conferenceID
        in: path
        description: The ID of the conference the track belongs to
        required: true
        schema:
          type: integer
          format: int64
      - name: trackID
        in: path
        description: The ID of the track for which the assignments should be finalized
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "201":
          description: Reviewers has been successfully finalized to this paper.
        "500":
          description: Internal Server Error. An unexpected server error occurred.
        "403":
          description: Forbidden. You are not allowed to finalize the assignments.
            You must be a chair of the track.
        "409":
          description: Conflict. There must be at least 3 reviewers assigned to each
            paper to finalize the assignments.
        "404":
          description: Not Found. The specified paper or user does not exist.
...
  /papers/{paperID}/reviews/phase:
    get:
      tags:
      - Reviews
      summary: Get the review phase of a paper
      description: Responds with how far along a paper is in the review process. Each
        paper whithin a track has its own phase. For more information check the description
        of the returned enum.
      operationId: getPhase
      parameters:
      - name: requesterID
        in: query
        description: The ID of a user making the request
        required: true
        schema:
          type: integer
          format: int64
      - name: paperID
        in: path
        description: The ID of the paper
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "500":
          description: Internal Server Error. An unexpected server error occurred.
        "403":
          description: "Forbidden. The requester must be a chair of the track the\
            \ paper is in, or a reviewer assigned to that paper."
        "404":
          description: Not Found. The requested paper was not found.
        "200":
          description: Successful retrieval of the paper review phase
          content:
            application/json:
              schema:
                type: string
                description: "Exactly how far along a paper is in the review process\n\
                  \nBEFORE_REVIEW: Before the Review phase of the track the paper\
                  \ is in\n\nIN_REVIEW: The reviewers all need to submit a review.\
                  \ Once that happens the discussion phase for the paper automatically\
                  \ starts.\n\nIN_DISCUSSION: The reviewers can now see each others\
                  \ reviews and comment on them. If all of the reviews are either\
                  \ positive or negative, the chair of the track can finalize the\
                  \ reviews which automatically puts the paper into the REVIEW phase.\n\
                  \nREVIEWED: The reviews cannot be changed now. Once all of the papers\
                  \ in a track are REVIEWED, the track goes into the final phase and\
                  \ the reviews can be viewed by the authors."
                enum:
                - BEFORE_REVIEW
                - IN_REVIEW
                - IN_DISCUSSION
                - REVIEWED
...
  /papers/{paperID}/reviewers:
    get:
      tags:
      - Reviews
      summary: Get the list of a reviewers for a given paper
      description: Responds with list of reviewers' IDs assigned to that paper.
      operationId: getReviewers
      parameters:
      - name: requesterID
        in: query
        description: The ID of the user making the request
        required: true
        schema:
          type: integer
          format: int64
      - name: paperID
        in: path
        description: The ID of the paper
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "500":
          description: Internal Server Error. An unexpected server error occurred.
        "200":
          description: Successful retrieval of the list of reviewers
          content:
            application/json:
              schema:
                type: array
                items:
                  type: integer
                  format: int64
        "403":
          description: "Forbidden. The requester must be a chair of the track that\
            \ the paper is in, or a reviewer also assigned to the given paper."
        "404":
          description: Not Found. The requested paper was not found.
...
  /conferences/{conferenceID}/tracks/{trackID}/phase:
    get:
      tags:
      - Conference Tracks
      summary: Get the review phase
      description: Responds with the review phase of the track. Consult with the documentation
        of the returned enum for more information.
      operationId: getPhase_1
      parameters:
      - name: requesterID
        in: query
        description: The ID of the user making the request.
        required: true
        schema:
          type: integer
          format: int64
      - name: conferenceID
        in: path
        description: The ID of the conference the track is in.
        required: true
        schema:
          type: integer
          format: int64
      - name: trackID
        in: path
        description: The ID of the track.
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "500":
          description: Internal Server Error. An unexpected server error occurred.
        "200":
          description: Successful pull of the review phase
          content:
            application/json:
              schema:
                type: string
                description: "Phases of the review process for a track.\n\nSUBMITTING:\
                  \ The review process hasn't started. Papers are still being submitted.\n\
                  \nBIDDING: Papers cannot be submitted now. Reviewers can bid on\
                  \ the papers, to indicate their expertise level. Ends after the\
                  \ bidding deadline for the track passes.\n\nASSIGNING: Chairs of\
                  \ the track can now assign papers to reviewers in the track.\n\n\
                  REVIEWING: Reviewing process. During this phase, the papers in the\
                  \ track are either in Review or in Discussion.\n\nFINAL: The reviews\
                  \ for all of the papers in the track have been finalized. They are\
                  \ now accessible to authors."
                enum:
                - SUBMITTING
                - BIDDING
                - ASSIGNING
                - REVIEWING
                - FINAL
        "403":
          description: "Forbidden. The requester must be a chair of the track, or\
            \ a reviewer in the track."
        "404":
          description: Not Found. The requested track or conference were not found.
...
components:
  schemas:
...
    DiscussionComment:
      type: object
      properties:
        commenterID:
          type: integer
          description: The ID of the commenter
          format: int64
          example: 1
        comment:
          type: string
          description: A confidential comment for other reviewers
          example: Some comment
      description: Discussion comment - it can only be seen by the reviewers and chairs
        and not authors.
...